# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  userId: string;
  /** @nullable */
  preferredName?: string | null;
  companionName: string;
  createdAt: string;
  updatedAt: string;
}

export interface ProfileUpdate {
  preferredName?: string;
  /** @minLength 1 */
  companionName?: string;
}

export interface Stats {
  conversationCount: number;
  messageCount: number;
  memoryCount: number;
  voiceProfileCount: number;
  /** @nullable */
  lastActiveAt: string | null;
}

export interface Conversation {
  id: number;
  /** @nullable */
  title: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  title?: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  /** @nullable */
  speakerName?: string | null;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface MessageExchange {
  userMessage: Message;
  assistantMessage: Message;
}

export interface VoiceMessageInput {
  /** Base64-encoded audio data */
  audioBase64: string;
  /** e.g. audio/webm */
  mimeType: string;
}

export interface VoiceMessageExchange {
  userMessage: Message;
  assistantMessage: Message;
  /** Base64-encoded synthesized speech audio of the reply */
  audioBase64: string;
  audioMimeType: string;
  /**
     * Name of the recognised speaker, or null if unrecognised
     * @nullable
     */
  speakerName?: string | null;
  /** True if a new voice profile was automatically created from a self-introduction */
  autoEnrolled?: boolean;
}

export interface Memory {
  id: number;
  content: string;
  /** @nullable */
  category: string | null;
  createdAt: string;
}

export interface VoiceProfile {
  id: number;
  name: string;
  /** @nullable */
  lastHeardAt?: string | null;
  createdAt: string;
}

export interface VoiceProfileInput {
  /** @minLength 1 */
  name: string;
  audioBase64: string;
  mimeType: string;
}

export interface VoiceProfileUpdate {
  /** @minLength 1 */
  name: string;
}