# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /voice-profiles:
    get:
      operationId: listVoiceProfiles
      tags: [voice-profiles]
      summary: List enrolled voice profiles for the current account
      responses:
        "200":
          description: List of voice profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VoiceProfile"
    post:
      operationId: enrollVoiceProfile
      tags: [voice-profiles]
      summary: Manually enroll a new voice profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceProfileInput"
      responses:
        "201":
          description: Enrolled voice profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceProfile"
  /voice-profiles/{id}:
    patch:
      operationId: updateVoiceProfile
      tags: [voice-profiles]
      summary: Rename a voice profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceProfileUpdate"
      responses:
        "200":
          description: Updated voice profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceProfile"
        "404":
          description: Not found
    delete:
      operationId: deleteVoiceProfile
      tags: [voice-profiles]
      summary: Remove an enrolled voice profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
components:
  schemas:
...
    VoiceProfile:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        lastHeardAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    VoiceProfileInput:
      type: object
      required: [name, audioBase64, mimeType]
      properties:
        name: { type: string, minLength: 1 }
        audioBase64: { type: string }
        mimeType: { type: string }
    VoiceProfileUpdate:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }