# lib/api-zod/src/generated/types/voiceProfile.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VoiceProfile {
  id: number;
  name: string;
  /** @nullable */
  lastHeardAt?: Date | null;
  createdAt: Date;
}