# lib/api-zod/src/generated/types/voiceProfileInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VoiceProfileInput {
  /** @minLength 1 */
  name: string;
  audioBase64: string;
  mimeType: string;
}